// insurise-front/insurise-front/src/app/main/apps/expert/models/SinisterNature.ts
export class sinisterNature{
    sinisterNatureId: number;
    title: string;
    code: string;
    branchId: number;
    idaEnabled: Boolean;
    transactionEnabled: Boolean;
    thirdPartyEnabled: Boolean;
    victimEnabled: Boolean;
    respEnabled: Boolean;
    respAutoEnabled: Boolean;
    affaireEnabled: Boolean;
    expertiseEnabled: Boolean;
    evalInterneEnabled: Boolean;
    evalInterneCorpEnabled: Boolean;
    regExpertEnabled: Boolean;
    isChecked: Boolean
} 